now handling mouse scroll and movement inputs

+++ engine/include/components/windowhandling/hwindow.h
@@ -39,7 +39,7 @@ namespace hf
 			std::string charData;
 			glm::ivec2 pointerPosition;
 			glm::ivec2 pointerDelta;
-			glm::ivec2 scrollDelta;
+			glm::vec2 scrollDelta;
 		};
 
 		std::string m_Title;
